x64 warnings were corrected. Because type 'size_t', used in project, is platform dependent explisit int64->int32 conversations were added to the code.

+++ Graphics/AbstractGraphicsEngine/AbstractGraphicsEngine.cpp
@@ -10,7 +10,7 @@ AbstractGraphicsEngine::~AbstractGraphicsEngine() {}
 GLFWwindow * AbstractGraphicsEngine::createWindow(char* title, size_t width, size_t height) {
 	setWindowSettings();
 
-	m_window = glfwCreateWindow(width, height, title, nullptr, nullptr);
+	m_window = glfwCreateWindow((int) width, (int) height, title, nullptr, nullptr);
 	if (!m_window)
 		throw Exceptions::WindowInitializationException();
 	m_window_width = width;
